refactor(web): migrate legacy pages to app directory (#cleanup)
Migrated the following pages from `web/src/pages/` to `web/src/app/` using the `page.tsx` convention:
- `SettingsPage.tsx` -> `app/settings/page.tsx`
- `LoginPage.tsx` -> `app/login/page.tsx`
- `AuthCallbackPage.tsx` -> `app/auth/callback/page.tsx`

Updated imports in `web/src/main.tsx` to reflect the new locations.
No logic changes were made to the components.
All tests passed.
