+++ web/src/main.tsx
@@ -23,7 +23,7 @@ import InsightsPage from '@/app/insights/page';
 import AuditPage from '@/app/audit/page';
 import HealthPage from '@/app/health/page';
 import SchedulesPage from '@/app/schedules/page';
-import SettingsPage from '@/pages/SettingsPage';
+import SettingsPage from '@/app/settings/page';
 import ToolsPage from '@/pages/ToolsPage';
 import MemoryExplorerPage from '@/pages/MemoryExplorerPage';
 import MemoryDetailPage from '@/pages/MemoryDetailPage';
@@ -31,8 +31,8 @@ import AgentMemoryGraphPage from '@/pages/AgentMemoryGraphPage';
 import ChannelsPage from '@/pages/ChannelsPage';
 import TemplatesPage from '@/pages/TemplatesPage';
 import ProvidersPage from '@/pages/ProvidersPage';
-import LoginPage from '@/pages/LoginPage';
-import AuthCallbackPage from '@/pages/AuthCallbackPage';
+import LoginPage from '@/app/login/page';
+import AuthCallbackPage from '@/app/auth/callback/page';
 import { ForbiddenView } from '@/views/ForbiddenView';
 import { NotFoundView } from '@/views/NotFoundView';
